requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
DateTime==4.3
Pillow==8.3.1
requests==2.24.0